+++ CHANGELOG.md
@@ -1,5 +1,10 @@
+### 0.1.7 - 1.20.1
+- fix spawnCloudEntity HelperMethod
+- Frostball now Deal Magic Damage on impact
+
 ### 0.1.6 - 1.20.1
 - created special Weapon Passive & Tooltip Methods to fix crashes with the Add-Ons
+- changed sword and axe attack damage attributes
 
 ### 0.1.5 - 1.20.1
 - add Witcher Drinks to Innkeeper Trading Pool